test.c
Hi guys this is the test code for the sample.s file and it works fine. I only defined the instructions that used in the .s file which means we need add more definition to it. any way I will show it to you guys tomorrow 
Have a great day